Fix shorten 64 to 32 warnings

+++ moor/archive_write_disk.hpp
@@ -62,7 +62,7 @@ namespace moor
         }
 
     private:
-        int writeDataBlock(const void* buf, size_t size, std::int64_t offset)
+        ssize_t writeDataBlock(const void* buf, size_t size, std::int64_t offset)
         {
             return archive_write_data_block(m_archive, buf, size, offset);
         }
